resolving errors after code refactoring and code refomatting
