--- quora-service/src/main/java/com/upgrad/quora/service/dao/UserAuthTokenDao.java
@@ -1,28 +1,0 @@
-package com.upgrad.quora.service.dao;
-
-import com.upgrad.quora.service.entity.UserAuthTokenEntity;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Repository;
-
-import javax.persistence.EntityManager;
-import javax.persistence.NoResultException;
-
-
-@Repository
-public class UserAuthTokenDao {
-    @Autowired
-    EntityManager entityManager;
-
-    /**
-     * retrieves the user auth token
-     * @param accessToken
-     * @return
-     */
-    public UserAuthTokenEntity getAuthToken(final String accessToken) {
-        try {
-            return entityManager.createNamedQuery("userAuthTokenByAccessToken", UserAuthTokenEntity.class).setParameter("accessToken", accessToken).getSingleResult();
-        } catch (NoResultException nre) {
-            return null;
-        }
-    }
-}
