+++ quora-service/src/main/java/com/upgrad/quora/service/dao/UserDao.java
@@ -3,8 +3,6 @@ package com.upgrad.quora.service.dao;
 import com.upgrad.quora.service.entity.UserAuthTokenEntity;
 import com.upgrad.quora.service.entity.UserEntity;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
 import javax.persistence.NoResultException;
@@ -15,12 +13,26 @@ public class UserDao {
 
     @PersistenceContext
     private EntityManager entityManager;
-    
+
     public UserEntity getUser(final String userUuid) {
         try {
             return entityManager.createNamedQuery("userByUuid", UserEntity.class).setParameter("uuid", userUuid)
                     .getSingleResult();
         } catch (NoResultException ex) {
+            return null;
+        }
+    }
+
+    /**
+     * retrieves the user auth token
+     *
+     * @param accessToken
+     * @return
+     */
+    public UserAuthTokenEntity getAuthToken(final String accessToken) {
+        try {
+            return entityManager.createNamedQuery("userAuthTokenByAccessToken", UserAuthTokenEntity.class).setParameter("accessToken", accessToken).getSingleResult();
+        } catch (NoResultException nre) {
             return null;
         }
     }
